requirements.txt
faiss-cpu>=1.7.4,<2.0
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.0.8
openai>=1.0.0
python-dotenv
sqlite-utils
streamlit
tiktoken
